refactor: Combine campaign segment into landing page
