--- components/CampaignSegment/index.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import CampaignCardsContainer from './FeaturedCampaignsContainer'
-import FeaturedCampaignsContainer from './CampaignCardsContainer'
-
-const CampaignSegment = () => {
-  return (
-    <>
-      <CampaignCardsContainer />
-      <FeaturedCampaignsContainer />
-    </>
-  )
-}
-
-export default CampaignSegment
