+++ components/LandingPage/CampaignSegment/CampaignCardsContainer/CampaignCard.tsx
