+++ components/LandingPage/CampaignSegment/CampaignCardsContainer/index.tsx
@@ -13,9 +13,9 @@ const CampaignCardsContainer = () => {
     >
       <div>Line</div>
       <Box display="flex" justifyContent="space-evenly" width="100%">
-        <CampaignCard imgSrc="/campaign_card_img_1.png" />
-        <CampaignCard imgSrc="/campaign_card_img_2.png" />
-        <CampaignCard imgSrc="/campaign_card_img_3.png" />
+        <CampaignCard imgSrc="/images/landing/campaign_card_img_1.png" />
+        <CampaignCard imgSrc="/images/landing/campaign_card_img_2.png" />
+        <CampaignCard imgSrc="/images/landing/campaign_card_img_3.png" />
       </Box>
       <Button variant="contained" size="large">
         See More
