+++ components/LandingPage/CampaignSegment/FeaturedCampaignsContainer/CampaignPanel.tsx
