+++ components/LandingPage/CampaignSegment/index.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import FeaturedCampaignsContainer from './FeaturedCampaignsContainer'
+import CampaignCardsContainer from './CampaignCardsContainer'
+
+const CampaignSegment = () => {
+  return (
+    <>
+      <FeaturedCampaignsContainer />
+      <CampaignCardsContainer />
+    </>
+  )
+}
+
+export default CampaignSegment
