+++ components/LandingPage/MainSegment/DescriptionContainer/StatsPanel.tsx
