+++ components/LandingPage/index.tsx
@@ -1,9 +1,11 @@
+import CampaignSegment from './CampaignSegment'
 import MainSegment from './MainSegment'
 
 const LandingPage = () => {
   return (
     <>
       <MainSegment />
+      <CampaignSegment />
     </>
   )
 }
